Deprecate static reflection
It has become a burden to maintain, for no benefit since we cannot find
usages for it in Doctrine itself. The plan is to create a new major
version without this types, so that php 8 compatibility is achieved.

After this, the package will consist in only 2 classes that should
probably moved elsewhere.